Pre_, at_ and post_gest_duration are double

+++ app/src/main/java/com/bailout/stickk/ubi4/models/DataCollectionSettings.kt
@@ -3,7 +3,7 @@ package com.bailout.stickk.ubi4.models
 data class DataCollectionSettings(
     val nCycles: Int,
     val baselineDuration: Int,
-    val preGestDuration: Int,
-    val atGestDuration: Int,
-    val postGestDuration: Int
+    val preGestDuration: Double,
+    val atGestDuration: Double,
+    val postGestDuration: Double
 )
